requirements.txt
# Web Framework
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21

# Database
sqlalchemy==2.0.46
asyncpg==0.31.0
psycopg2-binary==2.9.10
alembic
greenlet==3.3.0

# Pydantic / Settings
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.2.0

# Auth
python-jose==3.5.0
passlib==1.7.4
bcrypt==3.2.0

# HTTP Client
httpx==0.28.1

# AI / ML
openai==2.15.0
transformers==4.57.6
faiss-cpu==1.13.2
numpy==2.4.1
pillow==12.1.0

# Scheduler
apscheduler==3.11.2

# Utilities
python-dotenv==1.2.1
